reactive-flows/cnf-combustion/gnns/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
argcomplete==3.3.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorlog==6.8.2
cycler==0.12.1
distlib==0.3.8
docstring_parser==0.16
filelock==3.14.0
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
googledrivedownloader==0.4
grpcio==1.64.0
h5py==3.4.0
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonargparse==4.3.0
kaleido==0.2.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.4.3
multidict==6.0.5
networkx==2.6.3
nox==2024.4.15
numpy==1.21.2
oauthlib==3.2.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.3.1
protobuf==5.27.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyDeprecate==0.3.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.7
pytorch-ranger==0.1.1
pytz==2024.1
PyYAML==6.0
randomname==0.1.5
rdflib==7.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
setuptools==45.2.0
six==1.16.0
tenacity==8.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
torch==1.13.1+cu117
torch-geometric==2.0.3
torch-optimizer==0.1.0
torch-scatter==2.0.9
torch-sparse==0.6.17+pt113cu117
torchmetrics==0.5.1
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
Werkzeug==3.0.3
wheel==0.34.2
yacs==0.1.8
yarl==1.9.4
zipp==3.19.0